Adding null as allowed param type for where queries
It's possible in Doctrine to pass `null` as a param type in a where query (just the simple queries) and Doctrine will convert that to the proper SQL syntax for the underlying RDMS.

This doesn't work for the whereIn methods, as `null` would result in an empty array when it's casted, which Doctrine checks for and throws an exception to prevent "WHERE IN()" queries.

The `$params !== null` checks aren't necessary here as the `is_array` check won't pass if the value is null.