Appending orderby fields to select for mysql driver as well
Fixes #37.

As discussed in the ticket, mysql 5.7 with the only_full_group_by option enabled (which is the default) requires that any fields used in the order by are used in the select (when DISTINCT is used).

Doctrine already has this logic in place for other drivers (as it's standard SQL), so just adding mysql to that list.

Also noticed that there was still some extra whitespace in one of the files I touched (on blank lines), so added the php-cs-fixer rule for that and updated all files (can hide those changes by adding `w=1` to the github query string when viewing the PR's changes).